+++ packages/breadcrumb/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.1](https://github.com/zendeskgarden/react-containers/compare/@zendeskgarden/container-breadcrumb@0.2.0...@zendeskgarden/container-breadcrumb@0.2.1) (2019-05-14)
+
+**Note:** Version bump only for package @zendeskgarden/container-breadcrumb
+
+
+
+
+
 # [0.2.0](https://github.com/zendeskgarden/react-containers/compare/@zendeskgarden/container-breadcrumb@0.1.3...@zendeskgarden/container-breadcrumb@0.2.0) (2019-05-06)