+++ packages/pagination/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.4](https://github.com/zendeskgarden/react-containers/compare/@zendeskgarden/container-pagination@0.1.3...@zendeskgarden/container-pagination@0.1.4) (2019-05-14)
+
+**Note:** Version bump only for package @zendeskgarden/container-pagination
+
+
+
+
+
+## [0.1.3](https://github.com/zendeskgarden/react-containers/compare/@zendeskgarden/container-pagination@0.1.2...@zendeskgarden/container-pagination@0.1.3) (2019-05-13)
+
+
+### Bug Fixes
+
+* **selection|buttongroup|pagination:** Pass focusRef through as ref ([#29](https://github.com/zendeskgarden/react-containers/issues/29)) ([9eb0528](https://github.com/zendeskgarden/react-containers/commit/9eb0528))
+
+
+
+
+
 ## [0.1.2](https://github.com/zendeskgarden/react-containers/compare/@zendeskgarden/container-pagination@0.1.1...@zendeskgarden/container-pagination@0.1.2) (2019-05-07)
 
 **Note:** Version bump only for package @zendeskgarden/container-pagination